requirements.txt
functions-framework
requests
google-auth
google-cloud-bigquery